+++ domain/src/main/java/be/xplore/notifyme/services/IVenueService.java
@@ -2,14 +2,14 @@ package be.xplore.notifyme.services;
 
 import be.xplore.notifyme.domain.User;
 import be.xplore.notifyme.domain.Venue;
-import be.xplore.notifyme.dto.CreateVenueDto;
 import be.xplore.notifyme.dto.GetVenueDto;
 import java.security.Principal;
 import java.util.List;
 
 public interface IVenueService {
 
-  Venue createVenue(CreateVenueDto createVenueDto, Principal principal);
+  Venue createVenue(String name, String description, String streetAndNumber,
+                    String postalCode, String village, String country, Principal principal);
 
   Venue getVenue(long id);
