+++ domain/src/main/java/be/xplore/notifyme/services/implementations/LineService.java
@@ -2,7 +2,6 @@ package be.xplore.notifyme.services.implementations;
 
 import be.xplore.notifyme.domain.Line;
 import be.xplore.notifyme.domain.Team;
-import be.xplore.notifyme.dto.CreateLineDto;
 import be.xplore.notifyme.exception.CrudException;
 import be.xplore.notifyme.persistence.ILineRepo;
 import be.xplore.notifyme.services.ILineService;
